requirements.txt
streamlit
opencv-python-headless
numpy
roboflow
Pillow